+++ dflib-jdbc/src/main/java/com/nhl/dflib/jdbc/connector/JdbcConsumer.java
@@ -1,9 +1,0 @@
-package com.nhl.dflib.jdbc.connector;
-
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface JdbcConsumer<T> {
-
-    void accept(T t) throws SQLException;
-}