+++ dflib-jdbc/src/main/java/com/nhl/dflib/jdbc/connector/JdbcSupplier.java
@@ -1,9 +1,0 @@
-package com.nhl.dflib.jdbc.connector;
-
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface JdbcSupplier<T> {
-
-    T get() throws SQLException;
-}