+++ dflib-jdbc/src/main/java/com/nhl/dflib/jdbc/connector/StatementBinderFactory.java
@@ -12,5 +12,15 @@ public interface StatementBinderFactory {
 
     StatementBinder createBinder(PreparedStatement statement) throws SQLException;
 
-    StatementBinderFactory withFixedParams(DbColumnMetadata[] params);
+    /**
+     * Creates a derived binder factory for a known set of binding parameters described by {@link DbColumnMetadata}
+     * array argument. This only works for a known PreparedStatement structure and avoids the need to read
+     * parameters metadata from a PreparedStatement (as some DBs won't provide them correctly).
+     *
+     * @param paramsMetadata an array of parameter descriptors that must match the number and type of bindings in
+     *                       any PreparedStatement passed to {@link #createBinder(PreparedStatement)}.
+     * @return a new amended StatementBinderFactory
+     */
+    // TODO: dirty
+    StatementBinderFactory withFixedParams(DbColumnMetadata[] paramsMetadata);
 }