+++ dflib-jdbc/src/test/java/com/nhl/dflib/jdbc/connector/metadata/DbMetadataTest.java
@@ -12,7 +12,7 @@ public class DbMetadataTest {
     @Test
     public void testToCatalogSchemaName_CatalogNoSchema() {
 
-        DbMetadata md = new DbMetadata(DbType.MYSQL, mock(DatabaseMetaData.class));
+        DbMetadata md = new DbMetadata(DbFlavor.MYSQL, mock(DatabaseMetaData.class));
         assertArrayEquals(new String[]{"cat", null, "tab"}, md.toCatalogSchemaName("cat.tab"));
         assertArrayEquals(new String[]{null, null, "tab"}, md.toCatalogSchemaName("tab"));
         assertArrayEquals(new String[]{null, null, "cat.schema.tab"}, md.toCatalogSchemaName("cat.schema.tab"));